Added filtr and grades adding.

+++ Lab_10.App/MainWindow.xaml.cs
@@ -3,6 +3,8 @@ using Lab_10.Model;
 using Lab_10.Model.Attributes;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
+using System.Diagnostics;
 using System.Globalization;
 using System.Linq;
 using System.Reflection;
@@ -26,25 +28,28 @@ namespace Lab_10.App
     public partial class MainWindow : Window
     {
         #region Comments
-        public IList<Student> Students { get; set; }
+        //public IList<Student> Students { get; set; }
+        //private readonly GradesConverter gradesConverter = new GradesConverter();
 
-        public class GradesConverter : IValueConverter
-        {
-            public object? Convert(object value, Type targetType, object parametr, CultureInfo culture)
-            {
-                if (value is not List<Grade> grades) return null;
-                return string.Join(", ", grades.Select(g => $"{g.Subject}: {g.Value}"));
-            }
+        //public class GradesConverter : IValueConverter
+        //{
+        //    public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        //    {
+        //        if (value is not List<Grade> grades) return null;
 
-            public object? ConvertBack(object value, Type targetType, object parametr, CultureInfo culture)
-            {
-                throw new NotImplementedException();
-            }
-        }
+        //        return string.Join("; ", grades.Select(g => $"{g.Subject}: {g.Value}"));
+        //    }
+
+        //    public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        //    {
+        //        throw new NotImplementedException();
+        //    }
+        //}
         #endregion
 
         private readonly IRepository<Student> _studentRepository;
         private readonly IRepository<Grade> _gradeRepository;
+        private ICollectionView _collectionView;
 
         public MainWindow()
         {
@@ -53,74 +58,138 @@ namespace Lab_10.App
             _gradeRepository = new Repository<Grade>(connString);
             InitializeComponent();
             RefreshGrid();
+            _collectionView = CollectionViewSource.GetDefaultView(StudentsDg.ItemsSource);
 
-            #region Comments
-            //Students = new List<Student>
-            //{
-            //    new Student(){FirstName = "Jan", LastName="Kowalski", Faculty="WIMII", StudentNo = 116843, Grades = new List<Grade>(){
-            //        new Grade("PO",4.5),
-            //        new Grade("AM",3.5) }
-            //    },
-            //    new Student(){FirstName = "Michał", LastName="Nowak", Faculty="WIMII", StudentNo = 468735, Grades = new List<Grade>()
-            //    {
-            //        new Grade("PO",4.5),
-            //        new Grade("LM",5.0) }
-            //    },
-            //    new Student(){FirstName = "Marcin", LastName="Jakubski", Faculty="WIP", StudentNo = 647674, Grades = new List<Grade>()
-            //    {
-            //        new Grade("GH",2.0),
-            //        new Grade("WF",5.0) }
-            //    }
-            //};
-            //StudentsDg.Columns.Add(new DataGridTextColumn() { Header = "Imię", Binding = new Binding("FirstName") });
-            //StudentsDg.Columns.Add(new DataGridTextColumn() { Header = "Nazwisko", Binding = new Binding("LastName") });
-            //StudentsDg.Columns.Add(new DataGridTextColumn() { Header = "Wydział", Binding = new Binding("Faculty") });
-            //StudentsDg.Columns.Add(new DataGridTextColumn() { Header = "Nr albumu", Binding = new Binding("StudentNo") });
-            //StudentsDg.Columns.Add(new DataGridTextColumn() { Header = "Oceny", Binding = new Binding("Grades") { Converter = new GradesConverter() } });
-            //StudentsDg.AutoGenerateColumns = false;
-            //StudentsDg.ItemsSource = Students;
-            //StudentsDg.IsReadOnly = true;
-            //StudentsDg.IsReadOnly = true;
-            #endregion
+        #region Comments
+        //Students = new List<Student>
+        //{
+        //    new Student(){FirstName = "Jan", LastName="Kowalski", Faculty="WIMII", StudentNo = 116843, Grades = new List<Grade>(){
+        //        new Grade("PO",4.5),
+        //        new Grade("AM",3.5) }
+        //    },
+        //    new Student(){FirstName = "Michał", LastName="Nowak", Faculty="WIMII", StudentNo = 468735, Grades = new List<Grade>()
+        //    {
+        //        new Grade("PO",4.5),
+        //        new Grade("LM",5.0) }
+        //    },
+        //    new Student(){FirstName = "Marcin", LastName="Jakubski", Faculty="WIP", StudentNo = 647674, Grades = new List<Grade>()
+        //    {
+        //        new Grade("GH",2.0),
+        //        new Grade("WF",5.0) }
+        //    }
+        //};
+        //StudentsDg.Columns.Add(new DataGridTextColumn() { Header = "Imię", Binding = new Binding("FirstName") });
+        //StudentsDg.Columns.Add(new DataGridTextColumn() { Header = "Nazwisko", Binding = new Binding("LastName") });
+        //StudentsDg.Columns.Add(new DataGridTextColumn() { Header = "Wydział", Binding = new Binding("Faculty") });
+        //StudentsDg.Columns.Add(new DataGridTextColumn() { Header = "Nr albumu", Binding = new Binding("StudentNo") });
+        //StudentsDg.Columns.Add(new DataGridTextColumn() { Header = "Oceny", Binding = new Binding("Grades") { Converter = new GradesConverter() } });
+        //StudentsDg.AutoGenerateColumns = false;
+        //StudentsDg.ItemsSource = Students;
+        //StudentsDg.IsReadOnly = true;
+        //StudentsDg.IsReadOnly = true;
+        #endregion
         }
 
         private void RefreshGrid()
         {
             var studentList = _studentRepository.Select();
             foreach (var student in studentList)
+            {
                 student.Grades = _gradeRepository.Select(new Tuple<string, string, object, string>("StudentNo", "=", student.StudentNo, null));
+                //Debug.WriteLine($"Student: {student.FirstName} {student.LastName}, Grades Count: {student.Grades.Count}");
+                //Debug.WriteLine($"Student: {student.FirstName} {student.LastName}, Grades: {student.JoinedGrades}");
+                //student.JoinedGrades = gradesConverter.Convert(student.Grades, typeof(string), null, CultureInfo.InvariantCulture) as string;
+            }
             SetGrid(studentList);
+            ApplyDateOfBirthFilter();
         }
-
         private void SetGrid<T>(List<T> list) where T : new()
         {
             Type type = typeof(T);
-            if (type.GetCustomAttribute<DbTabAttribute>() == null)
-                return;
+            if (type.GetCustomAttribute<DbTabAttribute>() == null) { return; }
             StudentsDg.Columns.Clear();
             foreach (var prop in type.GetProperties())
             {
                 var col = prop.GetCustomAttribute<DisplayGridAttribute>();
                 if (col != null)
+                {
                     StudentsDg.Columns.Add(new DataGridTextColumn()
-                    { Header = col.Title ?? prop.Name, Binding = new Binding(prop.Name) });
+                    {
+                        Header = col.Title ?? prop.Name,
+                        Binding = new Binding(prop.Name)
+                    });
+                }
             }
             StudentsDg.ItemsSource = list;
             StudentsDg.AutoGenerateColumns = false;
             StudentsDg.Items.Refresh();
         }
 
+        #region Filtrowanie
 
+        private void ApplyDateOfBirthFilter()
+        {
+            if (_collectionView != null)
+            {
+                if (filterDatePck.SelectedDate.HasValue)
+                {
+                    DateTime selectedDate = filterDatePck.SelectedDate.Value.Date;
+                    FilterByDateOfBirth(selectedDate);
+                }
+                else
+                {
+                    ClearDateOfBirthFilter();
+                }
+            }
+        }
+
+        private void FilterByDateOfBirth(DateTime selectedDate)
+        {
+            _collectionView.Filter = item =>
+            {
+                if (item is Student student)
+                {
+                    return student.DateOfBirth.Date == selectedDate.Date;
+                }
+
+                return false;
+            };
+        }
+
+        private void ClearDateOfBirthFilter()
+        {
+            if (_collectionView != null)
+            {
+                _collectionView.Filter = null;
+            }
+        }
+
+        private void filterBtn_Click(object sender, RoutedEventArgs e)
+        {
+            if (filterDatePck.SelectedDate.HasValue)
+            {
+                DateTime selectedDate = filterDatePck.SelectedDate.Value.Date;
+                FilterByDateOfBirth(selectedDate);
+            }
+            else
+            {
+                ClearDateOfBirthFilter();
+            }
+        }
+
+        #endregion
 
         private void AddMi_Click(object sender, RoutedEventArgs e)
         {
             AddStudentWindow addStudentWindow;
             if (StudentsDg.SelectedItem != null && StudentsDg.SelectedItem is Student student)
+            {
                 addStudentWindow = new AddStudentWindow(student);
+            }
             else
+            {
                 addStudentWindow = new AddStudentWindow();
-            //if (addStudentWindow.ShowDialog() == true)
-            //    RefreshGrid();
+            }
 
             if (addStudentWindow.ShowDialog() == true)
             {
@@ -128,40 +197,50 @@ namespace Lab_10.App
 
                 if (StudentsDg.SelectedItem != null)
                 {
-                    // Update existing student in the repository
                     _studentRepository.Update(updatedStudent);
                 }
                 else
                 {
-                    // Add the new student to the repository
                     _studentRepository.Insert(updatedStudent);
                 }
-
-                // Refresh the grid after updating or adding a student
                 RefreshGrid();
+                ApplyDateOfBirthFilter();
             }
-
-            //var dodajStudenta = new AddStudentWindow();
-            //if (dodajStudenta.ShowDialog() == true)
-            //{
-            //    Students.Add(dodajStudenta.Student);
-            //    StudentsDg.Items.Refresh();
-            //}
         }
 
         private void RemoveMi_Click(object sender, RoutedEventArgs e)
         {
             if (StudentsDg.SelectedItem is Student studentToRemove)
             {
-                _studentRepository.Delete(studentToRemove);
+                string confirmationMessage = studentToRemove.HasGrades() ? "Ten student ma przypisane oceny. Czy chcesz usunąć również oceny?" : "Czy na pewno chcesz usunąć tego studenta?";
+
+                MessageBoxResult result = MessageBox.Show(confirmationMessage, "Potwierdzenie", MessageBoxButton.YesNo, MessageBoxImage.Question);
+
+                if (result == MessageBoxResult.Yes)
+                {
+                    if (studentToRemove.HasGrades())
+                    {
+                        // Usuń oceny studenta
+                        foreach (var grade in studentToRemove.Grades.ToList())
+                        {
+                            _gradeRepository.Delete(grade);
+                        }
+                    }
+
+                    _studentRepository.Delete(studentToRemove);
+                    RefreshGrid();
+                    ApplyDateOfBirthFilter();
+                }
+            }
+        }
+        private void AddGradeMi_Click(object sender, RoutedEventArgs e)
+        {
+            if (StudentsDg.SelectedItem is Student selectedStudent)
+            {
+                AddGradeWindow addGradeWindow = new AddGradeWindow(selectedStudent, _gradeRepository);
+                addGradeWindow.ShowDialog();
                 RefreshGrid();
             }
-
-            //if (StudentsDg.SelectedItem is Student)
-            //{
-            //    Students.Remove((Student)StudentsDg.SelectedItem);
-            //    StudentsDg.Items.Refresh();
-            //}
         }
 
         private void exitMi_Click(object sender, RoutedEventArgs e)
